moved utils into vis, added support for node positions in plot
